Optimize playback
Previously, we used an <audio> element to play the audio.
This was clunky, since the playback range timing was imprecise.
It was also not very efficient, since the browser had to load the
object url every time the audio was played (which was super wasteful,
since we already had loaded the `AudioBuffer`!).

This commit uses `AudioContext.createBufferSource` to play the audio.
This is a major improvement, since it allows us to precisely control
the playback range, while also simultaneously eliminating the need to
load the object url.